pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "google-generativeai>=0.8.6",
    "gunicorn>=25.0.2",
    "psycopg2-binary>=2.9.11",
]
